finished css styling. Added code so that the user who creates the suggestion can update or delete it

+++ controllers/suggestionController.js
@@ -105,12 +105,14 @@ router.route('/suggestion/update/:id').get(async function (req, res, next) {
 })
 
 router.route('/suggestion/:category').get(async function (req, res) {
+
   const suggestions = await Suggestion.find()
   const chosenCategory = req.params.category
   res.render('index.ejs',
     {
       suggestions: suggestions,
-      chosenCategory: chosenCategory
+      chosenCategory: chosenCategory,
+
     }
   )
 })
@@ -132,7 +134,25 @@ router.route('/suggestion/:category').get(async function (req, res) {
 
 router.route('/suggestion/:id').delete(async function (req, res) {
   console.log('RUNNING THE POST EJS CODE')
+
+  if (!req.session.user) {
+    return res.status(402).send({ message: "you must be logged in to delete a suggestion" })
+  }
   const suggestionId = req.params.id
+  const suggestion = await Suggestion.findById(suggestionId).populate('user')
+
+  console.log(`req params id ${req.params.id}`)
+  console.log(`SUGGESTION ${suggestion}`)
+  if (!suggestion.user) {
+    return res.status(404).send({ message: "Associated user not found for this suggestion." });
+  }
+  console.log(suggestion.user._id)
+  if (!suggestion.user._id.equals(req.session.user._id)) {
+    return res.status(402).send({ message: "this is not your suggestion to delete!" })
+  }
+  if (!suggestion) {
+    return res.send({ message: "Suggestion doesn't exist." })
+  }
 
   await Suggestion.findByIdAndDelete(suggestionId)
   res.redirect('/')
@@ -140,17 +160,15 @@ router.route('/suggestion/:id').delete(async function (req, res) {
 
 
 router.route('/suggestion/update/:id').put(async function (req, res) {
-  // if(!req.session.user){
-  //   return res.status(402).send({message: "you must be logged in to update a mode of transport"})
-  //  }
+  if (!req.session.user) {
+    return res.status(402).send({ message: "you must be logged in to update a suggestion" })
+  }
   console.log("RUNNING CODE FROM CONTROLLER FOR PUT😄")
   const suggestionId = req.params.id
 
   const suggestion = await Suggestion.findById(suggestionId)
 
-  // if (!transport.user._id.equals(req.session.user._id)) {
-  //   return res.status(402).send({ message: "this is not your destination to update!"})
-  // }
+  
   const updateSuggestion = await Suggestion.findByIdAndUpdate(suggestionId, req.body, { new: true })
   if (!updateSuggestion) {
     return res.status(404).send({ message: "Suggestion not found" });
@@ -161,11 +179,25 @@ router.route('/suggestion/update/:id').put(async function (req, res) {
 
 router.route('/suggestion/new').post(async function (req, res) {
   console.log("RUNNING CODE FROM CONTROLLER")
-  // Get the new destination from the body of request
-  const newSuggestion = await Suggestion.create(req.body)
-  // Send back our destination with appropriate status code.
-  res.redirect('/')
+  try {
+
+    if (!req.session.user) {
+      return res.status(402).send({ message: "you must be logged in to save a suggestion" })
+    }
+
+    req.body.user = req.session.user
+    console.log(req.body)
+    // Get the new destination from the body of request
+    const newSuggestion = await Suggestion.create(req.body)
+    // Send back our destination with appropriate status code.
+    res.redirect('/')
+  } catch (e) {
+    next(e)
+  }
 })
+
+
+
